requirements.txt
# MetaFinder Dependencies

# Core metadata extraction
PyExifTool>=0.5.6

# Image processing
Pillow>=10.0.0

# Database (built-in, but listing for documentation)
# sqlite3 is included with Python

# Optional: Enhanced parsing
python-docx>=1.0.0
openpyxl>=3.1.0

# Geocoding for GPS coordinates
geopy>=2.4.0

# Progress bars for CLI
tqdm>=4.66.0

# Date parsing
python-dateutil>=2.8.0
